// src/entity/Profile/model/types/profile.ts
import { Country } from 'entity/Country';
import { Currency } from 'entity/Currency';

export interface Profile {
    id?: string;
    first?: string;
    lastname?: string;
    age?: number;
    currency?: Currency;
    country?: Country;
    city?: string;
    username?: string;
    avatar?: string;
}